--- docs/.vuepress/theme/components/Home.vue
@@ -1,179 +1,0 @@
-<template>
-  <main class="home" :aria-labelledby="heroText ? 'main-title' : undefined">
-    <header class="hero">
-      <svg class="logo" :alt="heroAlt"  version="1.1" id="Ebene_1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px"
-           viewBox="0 0 492 110.7" enable-background="new 0 0 492 110.7" xml:space="preserve">
-<g>
-	<path class="light" fill="#009FE3" d="M26.9,107c1.9,0.3,3.9,0.5,5.7,0.6c1.9,0.1,3.7,0.1,5.3,0.1c1.6,0,3.3-0.1,5.1-0.2c1.7-0.1,3.5-0.3,5.2-0.5
-		V20.5h26c0.6-2.9,0.9-6,0.9-9.2c0-3-0.3-6-0.9-8.9H0.9C0.3,5.4,0,8.3,0,11.3c0,3.2,0.3,6.3,0.9,9.2h26V107L26.9,107z M80.9,107.2
-		c1.7,0.3,3.5,0.5,5.2,0.6c1.7,0.1,3.4,0.1,5,0.1c1.6,0,3.3,0,5.1-0.1c1.7-0.1,3.5-0.3,5.2-0.6V67.6c0-4.1,0.5-7.4,1.4-9.9
-		c0.9-2.5,2.1-4.5,3.5-5.9c1.4-1.4,3-2.4,4.7-2.8c1.7-0.5,3.4-0.7,5.1-0.7c4,0,6.8,1.4,8.4,4.2c1.6,2.8,2.5,6.8,2.5,12.1v42.6
-		c1.7,0.3,3.5,0.5,5.2,0.6c1.7,0.1,3.4,0.1,5.1,0.1c1.6,0,3.3,0,5.1-0.1c1.7-0.1,3.5-0.3,5.2-0.6V59c0-10-2.2-17.3-6.5-22
-		c-4.3-4.7-10.6-7-19-7c-3,0-5.7,0.4-8,1.2c-2.3,0.8-4.3,1.7-6,2.8c-1.6,1.1-3,2.3-4.1,3.5c-1.1,1.2-2,2.3-2.7,3.1V2
-		c-1.7-0.3-3.5-0.5-5.2-0.6c-1.7-0.1-3.4-0.1-5.1-0.1c-1.7,0-3.3,0-5.1,0.1c-1.7,0.1-3.4,0.3-5.1,0.6V107.2L80.9,107.2z M337.8,31.6
-		c-1.7-0.5-3.6-0.8-5.5-0.9c-1.9-0.1-3.7-0.2-5.5-0.2c-5.8,0-11.1,1-15.9,3.1c-4.8,2.1-8.8,4.9-12.1,8.5c-3.3,3.6-5.9,7.8-7.7,12.7
-		c-1.8,4.9-2.7,10.1-2.7,15.6c0,7,1,12.9,3.1,17.7c2.1,4.8,4.9,8.8,8.5,11.9c3.6,3.1,7.7,5.3,12.4,6.7c4.7,1.4,9.6,2.1,14.8,2.1
-		c6.7,0,12.4-0.4,17.1-1.1c4.7-0.7,9.3-1.8,13.9-3.1V0.7c-1.7-0.3-3.5-0.5-5.2-0.6C351.3,0,349.6,0,348,0c-1.6,0-3.3,0-5,0.1
-		c-1.7,0.1-3.4,0.3-5.2,0.6V31.6L337.8,31.6z M337.8,90.6c-1.6,0.5-3.2,0.8-4.5,0.9c-1.4,0.1-2.9,0.2-4.5,0.2
-		c-2.7,0-5.2-0.4-7.6-1.1c-2.3-0.7-4.3-1.9-6-3.6c-1.7-1.6-3-3.9-4-6.6c-1-2.8-1.5-6.1-1.5-10.1c0-3.2,0.4-6.3,1.1-9.2
-		c0.7-3,1.9-5.5,3.4-7.7c1.6-2.2,3.5-3.9,5.9-5.2c2.4-1.3,5.2-2,8.4-2c1.8,0,3.4,0.1,4.6,0.2c1.2,0.1,2.8,0.5,4.7,0.9V90.6
-		L337.8,90.6z M435.2,75.8c0.2-1.1,0.4-2.5,0.5-4.3c0.1-1.8,0.2-3.6,0.2-5.5c0-5-0.7-9.8-2.2-14.2c-1.5-4.4-3.6-8.2-6.5-11.4
-		c-2.9-3.2-6.3-5.7-10.4-7.6c-4.1-1.8-8.7-2.8-14-2.8c-6,0-11.3,1-15.8,3.1c-4.5,2.1-8.3,4.9-11.3,8.5c-3.1,3.6-5.4,7.8-6.9,12.7
-		c-1.6,4.9-2.3,10-2.3,15.6c0,5.6,0.8,10.8,2.3,15.6c1.5,4.8,3.8,8.8,7,12.3c3.1,3.4,7.2,6.1,12.2,8.1c5,1.9,13.8,2.9,20.8,2.9
-		c9.5,0,15.2-1.5,22.8-4.4c-0.1-3.1-0.5-6-1.3-8.7c-0.8-2.7-1.7-5-2.8-7c-3.1,1.2-6.4,2.1-9.8,2.7c-3.4,0.6-6.8,0.9-10.1,0.9
-		c-6,0-10.9-1.3-14.5-3.9c-3.7-2.6-5.7-6.8-6.1-12.7H435.2L435.2,75.8z M387.1,61.4c0.5-4.9,2-8.8,4.5-11.5
-		c2.5-2.7,6.2-4.1,11.1-4.1c4.3,0,7.6,1.4,10,4.3c2.4,2.9,3.7,6.6,3.9,11.3H387.1L387.1,61.4z M446.6,107.2c1.8,0.3,3.6,0.5,5.3,0.6
-		c1.7,0.1,3.4,0.1,5,0.1c1.6,0,3.3,0,5.1-0.1c1.7-0.1,3.5-0.3,5.2-0.6V71.4c0-5.2,0.6-9.3,1.8-12.2c1.2-2.9,2.7-5.1,4.6-6.5
-		c1.8-1.5,3.8-2.4,5.8-2.7c2-0.3,3.9-0.5,5.5-0.5c0.3,0,1.1,0,2.4,0c1.3,0,2.5,0.1,3.6,0.3c0.4-1.7,0.7-3.6,0.8-5.5
-		c0.1-1.9,0.2-3.7,0.2-5.4c0-1.4,0-2.7-0.1-3.9c-0.1-1.3-0.2-2.4-0.4-3.3c-0.6-0.2-1.5-0.3-2.8-0.4c-1.3-0.1-2.4-0.1-3.3-0.1
-		c-4.7,0-8.5,1.2-11.6,3.6c-3.1,2.4-5.7,5.1-7.9,8.1c-0.1-0.7-0.2-1.5-0.4-2.5c-0.1-1-0.3-2-0.5-3.1c-0.2-1-0.4-2-0.7-2.9
-		c-0.2-0.9-0.5-1.6-0.8-2.1c-1.4-0.3-2.7-0.5-4.1-0.7c-1.4-0.2-2.8-0.3-4.3-0.3c-1.6,0-3,0.1-4.3,0.2s-2.7,0.4-4.1,0.7V107.2
-		L446.6,107.2z"/>
-  <path class="dark" fill="#A2DAF8" d="M244.6,110.5c-20.6,0-30.4-12.3-32.5-17.1c-1.8-4.4-2.8-10-3-17l0,0v-2c0-0.1,0-0.3,0-0.4v-7
-		c0-3.8-0.3-7-0.9-9.5c-0.6-2.6-1.6-4.6-2.9-6.1c-1.3-1.5-2.9-2.6-4.9-3.2c-1.9-0.6-4.2-0.9-6.8-0.9c-2.6,0-4.9,0.3-6.8,0.9
-		c-1.9,0.6-3.6,1.7-4.9,3.2c-1.3,1.5-2.3,3.5-2.9,6.1c-0.6,2.6-0.9,5.8-0.9,9.8v39.8c-0.1,0-0.2,0-0.3,0c-3.4,0-6.5-8.3-9.8-8.3
-		c-3.4,0-7-22.9-10.5-23.5v-8.6c0-8,1-14.5,3.1-19.3c2-4.8,10.4-8.5,13.9-11c3.4-2.5,6.9-6.1,19.9-6.1c20.6,0,30.4,12.3,32.5,17.1
-		c1.8,4.4,2.8,10,3,17l0,0v2c0,0.1,0,0.3,0,0.4v7c0,3.8,0.3,7,0.9,9.5c0.6,2.6,1.6,4.6,2.9,6.1c1.3,1.5,2.9,2.6,4.9,3.2
-		c1.9,0.6,4.2,0.9,6.8,0.9c2.6,0,4.9-0.3,6.8-0.9c1.9-0.6,3.6-1.7,4.9-3.2c1.3-1.5,2.3-3.5,2.9-6.1c0.6-2.6,0.9-5.8,0.9-9.8V33.7
-		c0.1,0,0.2,0,0.3,0c3.4,0,6.5,8.3,9.8,8.3c3.4,0,7,22.9,10.5,23.5v8.6c0,8-1,14.5-3.1,19.3c-2,4.8-10.4,8.5-13.9,11
-		C261,107,257.6,110.5,244.6,110.5"/>
-  <path class="light" fill="#009FE3" d="M281.4,107V66.6c0-8-1-14.5-3.1-19.3c-2-4.8-4.8-8.5-8.2-11c-3.4-2.5-7.4-4.2-12-4.9
-		c-4.1-0.7-8.5-1-12.9-1.1c-4.5,0.1-8.8,0.4-12.9,1.1c-4.6,0.8-8.6,2.4-12,4.9c-3.4,2.5-6.2,6.2-8.2,11c-2,4.8-3.1,11.3-3.1,19.3V74
-		c0,3.8-0.3,7-0.9,9.5c-0.6,2.6-1.6,4.6-2.9,6.1c-1.3,1.5-2.9,2.6-4.9,3.2c-1.9,0.6-4.2,0.9-6.8,0.9c-2.6,0-4.9-0.3-6.8-0.9
-		c-1.9-0.6-3.6-1.7-4.9-3.2c-1.3-1.5-2.3-3.5-2.9-6.1c-0.6-2.6-0.9-5.8-0.9-9.8V33.9c-3.5-0.6-6.9-0.9-10.3-0.9
-		c-3.4,0-6.8,0.3-10.3,0.9v40.4c0,8,1,14.5,3.1,19.3c2,4.8,4.8,8.5,8.2,11c3.4,2.5,7.4,4.2,12,4.9c4.1,0.7,8.5,1,12.9,1.1
-		c4.5-0.1,8.8-0.4,12.9-1.1c4.6-0.8,8.6-2.4,12-4.9c3.4-2.5,6.2-6.2,8.2-11c1.9-4.4,2.9-10.3,3-17.4l0,0v-1.5c0-0.1,0-0.3,0-0.4l0,0
-		v-7.1c0-4,0.3-7.2,0.9-9.8c0.6-2.6,1.6-4.6,2.9-6.1c1.3-1.5,2.9-2.6,4.9-3.2c1.9-0.6,4.2-0.9,6.8-0.9c2.6,0,4.9,0.3,6.8,0.9
-		c1.9,0.6,3.6,1.7,4.9,3.2c1.3,1.5,2.3,3.5,2.9,6.1c0.6,2.6,0.9,5.8,0.9,9.8V107c3.5,0.6,6.9,0.9,10.3,0.9
-		C274.5,107.9,277.9,107.6,281.4,107L281.4,107z"/>
-</g>
-</svg>
-
-      <h1 v-if="heroText" id="main-title">
-        {{ heroText }}
-      </h1>
-
-      <p v-if="tagline" class="description">
-        {{ tagline }}
-      </p>
-
-      <p v-if="actions.length" class="actions">
-        <NavLink
-          v-for="action in actions"
-          :key="action.text"
-          class="action-button"
-          :class="[action.type]"
-          :item="action"
-        />
-      </p>
-    </header>
-
-    <div v-if="features.length" class="features">
-      <div v-for="feature in features" :key="feature.title" class="feature">
-        <h2>{{ feature.title }}</h2>
-        <p>{{ feature.details }}</p>
-      </div>
-    </div>
-
-    <div class="theme-default-content custom">
-      <Content />
-    </div>
-
-    <template v-if="footer">
-      <!-- eslint-disable-next-line vue/no-v-html -->
-      <div v-if="footerHtml" class="footer" v-html="footer" />
-      <div v-else class="footer" v-text="footer" />
-    </template>
-  </main>
-</template>
-
-<script lang="ts">
-import { computed, defineComponent } from 'vue'
-import {
-  usePageFrontmatter,
-  useSiteLocaleData,
-  withBase,
-} from '@vuepress/client'
-import { isArray } from '@vuepress/shared'
-import type { DefaultThemeHomePageFrontmatter } from '@vuepress/theme-default/lib/shared'
-import NavLink from '@vuepress/theme-default/lib/client/components/NavLink.vue'
-
-export default defineComponent({
-  name: 'Home',
-
-  components: {
-    NavLink,
-  },
-
-  setup() {
-    const frontmatter = usePageFrontmatter<DefaultThemeHomePageFrontmatter>()
-    const siteLocale = useSiteLocaleData()
-
-    const heroImage = computed(() => {
-      if (!frontmatter.value.heroImage) {
-        return null
-      }
-
-      return withBase(frontmatter.value.heroImage)
-    })
-
-    const heroText = computed(() => {
-      if (frontmatter.value.heroText === null) {
-        return null
-      }
-      return frontmatter.value.heroText || siteLocale.value.title || 'Hello'
-    })
-
-    const heroAlt = computed(
-      () => frontmatter.value.heroAlt || heroText.value || 'hero'
-    )
-
-    const tagline = computed(() => {
-      if (frontmatter.value.tagline === null) {
-        return null
-      }
-      return (
-        frontmatter.value.tagline ||
-        siteLocale.value.description ||
-        'Welcome to your VuePress site'
-      )
-    })
-
-    const actions = computed(() => {
-      if (!isArray(frontmatter.value.actions)) {
-        return []
-      }
-
-      return frontmatter.value.actions.map(
-        ({ text, link, type = 'primary' }) => ({
-          text,
-          link,
-          type,
-        })
-      )
-    })
-
-    const features = computed(() => {
-      if (isArray(frontmatter.value.features)) {
-        return frontmatter.value.features
-      }
-      return []
-    })
-
-    const footer = computed(() => frontmatter.value.footer)
-
-    const footerHtml = computed(() => frontmatter.value.footerHtml)
-
-    return {
-      heroImage,
-      heroAlt,
-      heroText,
-      tagline,
-      actions,
-      features,
-      footer,
-      footerHtml,
-    }
-  },
-})
-</script>
