+++ docs/.vuepress/theme/components/NavbarBrand.vue
@@ -1,0 +1,43 @@
+<script setup lang="ts">
+import {
+  ClientOnly,
+  useRouteLocale,
+  useSiteLocaleData,
+  withBase,
+} from '@vuepress/client'
+import { computed, h } from 'vue'
+import type { FunctionalComponent } from 'vue'
+import { useDarkMode, useThemeLocaleData } from '@vuepress/theme-default/lib/client/composables'
+import Logo from './Logo.vue'
+
+const routeLocale = useRouteLocale()
+const siteLocale = useSiteLocaleData()
+const themeLocale = useThemeLocaleData()
+const isDarkMode = useDarkMode()
+
+const navbarBrandLink = computed(
+  () => themeLocale.value.home || routeLocale.value
+)
+const navbarBrandTitle = computed(() => siteLocale.value.title)
+const navbarBrandLogo = computed(() => {
+  if (isDarkMode.value && themeLocale.value.logoDark !== undefined) {
+    return themeLocale.value.logoDark
+  }
+  return themeLocale.value.logo
+})
+
+</script>
+
+<template>
+  <RouterLink :to="navbarBrandLink">
+    <Logo />
+
+    <span
+      v-if="navbarBrandTitle"
+      class="site-name"
+      :class="{ 'can-hide': navbarBrandLogo }"
+    >
+      {{ navbarBrandTitle }}
+    </span>
+  </RouterLink>
+</template>
