+++ docs/.vuepress/theme/index.js
@@ -3,7 +3,8 @@ const { path } = require('@vuepress/utils')
 module.exports = {
   name: 'vuepress-theme-local',
   extends: '@vuepress/theme-default',
-  layouts: {
-    Layout: path.resolve(__dirname, 'layouts/Layout.vue'),
+  alias: {
+    '@theme/HomeHero.vue': path.resolve(__dirname, './components/HomeHero.vue'),
+    '@theme/NavbarBrand.vue': path.resolve(__dirname, './components/NavbarBrand.vue'),
   },
 }
