+++ package.json
@@ -13,14 +13,14 @@
   },
   "homepage": "https://github.com/thunder/thunder-distribution#readme",
   "devDependencies": {
-    "@vuepress/plugin-search": "^2.0.0-beta.27",
+    "@vuepress/plugin-search": "^2.0.0-beta.33",
     "eslint": "^7.32.0",
     "eslint-config-drupal": "^5.0.2",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "prettier": "^2.5.0",
+    "prettier": "^2.5.1",
     "request": "^2.88.2",
     "request-promise": "^4.2.6",
-    "vuepress": "^2.0.0-beta.27"
+    "vuepress": "^2.0.0-beta.33"
   },
   "dependencies": {},
   "scripts": {
